zmena long na unsigned long long pre pocet kostier

+++ VypocetKostier.cpp
@@ -32,7 +32,7 @@ double VypocetKostier::maxH(const std::vector<std::vector<double>>& hodnoty) {
 }
 
 
-long VypocetKostier::gauss(Matica& matica, int n) {
+long long VypocetKostier::gauss(Matica& matica, int n) {
     int r1 = 0;
     int s1 = 0;
     int nas = 0;
@@ -78,11 +78,11 @@ long VypocetKostier::gauss(Matica& matica, int n) {
 
 }
 
-long VypocetKostier::kofaktor(Matica& matica, int n, int r, int s) {
-    return (long) (pow(-1, r+s) * VypocetKostier::gauss(matica, n-1));
+unsigned long long VypocetKostier::kofaktor(Matica& matica, int n, int r, int s) {
+    return (unsigned long long) (pow(-1, r+s) * VypocetKostier::gauss(matica, n-1));
 }
 
-long VypocetKostier::celkovyVypocet(const Hrany& hrany, int reg, int n) {
+unsigned long long VypocetKostier::celkovyVypocet(const Hrany& hrany, int reg, int n) {
     std::vector<std::vector<double>> matica(n-1);
     for (int i= 0; i < n-1; i++) {
         std::vector<double> riadok(n-1);
